chatbot-docs/requirements.txt
streamlit
PyPDF2
langchain
langchain-community
sqlalchemy
psycopg2-binary
python-dotenv
